Updated taxon annotations data service to use Solr query. Partially updated annotations and taxa data service to use Solr counts (still queries twice).

+++ src/org/phenoscape/ws/resource/DistinctPhenotypesResource.java
@@ -58,7 +58,7 @@ public class DistinctPhenotypesResource extends AnnotationQueryingResource<Pheno
     }
 
     @Override
-    protected int queryForItemsCount(AnnotationsQueryConfig config) throws SQLException {
+    protected long queryForItemsCount(AnnotationsQueryConfig config) throws SQLException {
         return this.getDataStore().getCountOfDistinctPhenotypes(config);
     }
